requirements.txt
pytz
discord.py
python-dotenv
discord-components
PyNaCl